pyproject.toml
[project]
name = "agents-orchestration"
version = "0.1.0"
description = "A hierarchical agent orchestration project using LangGraph and uv."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ddgs>=9.12.1",
    "duckduckgo-search>=8.1.1",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "pandas>=3.0.2",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
]
